Refactor Scene into Importer
This was a major overhaul of the system. I got rid of `file` and `obj`
modules, refactored Importer and made tests mode granular.
