+++ src/lib.rs
@@ -1,5 +1,3 @@
-pub mod file;
-pub mod obj;
+pub mod importer;
 pub mod primitives;
-pub mod scene;
 pub mod viewport_renderer;
